user: RequestAndVerify
