+++ testdata/user.spew
@@ -1,6 +1,6 @@
-/user/alice@twitter {"kid":"kex1a4yj333g68pvd6hfqvufqkv4vy54jfe6t33ljd3kc9rpfty8xlgs2u3qxr","result":{"status":"ok","ts":1234567890050,"user":{"k":"kex1a4yj333g68pvd6hfqvufqkv4vy54jfe6t33ljd3kc9rpfty8xlgs2u3qxr","n":"alice","sq":1,"sr":"twitter","u":"https://twitter.com/alice/status/1"},"vts":1234567890051}}
-/user/name10@github {"kid":"kex1gwnjuu2yq9mzmantdrpxm77ly6p24myly36wefrp8epy5ra6l57qj97xgz","result":{"status":"ok","ts":1234567890005,"user":{"k":"kex1gwnjuu2yq9mzmantdrpxm77ly6p24myly36wefrp8epy5ra6l57qj97xgz","n":"name10","sq":1,"sr":"github","u":"https://gist.github.com/name10/1"},"vts":1234567890006}}
-/user/name11@github {"kid":"kex1v6l8uvev0fznxv4an5987lds2h6utmc6q6k6vmvckw0mdqgvguaq4850kc","result":{"status":"ok","ts":1234567890011,"user":{"k":"kex1v6l8uvev0fznxv4an5987lds2h6utmc6q6k6vmvckw0mdqgvguaq4850kc","n":"name11","sq":1,"sr":"github","u":"https://gist.github.com/name11/1"},"vts":1234567890012}}
-/user/name12@github {"kid":"kex1pdgn4kd5jfqptjsfqtks0yzy6wk9m0kzxphsd9yvzrdgadhrnuksyklezm","result":{"status":"ok","ts":1234567890017,"user":{"k":"kex1pdgn4kd5jfqptjsfqtks0yzy6wk9m0kzxphsd9yvzrdgadhrnuksyklezm","n":"name12","sq":1,"sr":"github","u":"https://gist.github.com/name12/1"},"vts":1234567890018}}
-/user/name13@github {"kid":"kex1jx3g5zm58q2e8fxeg62hjgyfy6hu3tvzezpekajyxkdeaw56fvaq46atcz","result":{"status":"ok","ts":1234567890023,"user":{"k":"kex1jx3g5zm58q2e8fxeg62hjgyfy6hu3tvzezpekajyxkdeaw56fvaq46atcz","n":"name13","sq":1,"sr":"github","u":"https://gist.github.com/name13/1"},"vts":1234567890024}}
-/user/name14@github {"kid":"kex1p0h0t20x08n28cf5lcncx7llxtrukh6agn4qn09su4pt444ycrxq2x9f8c","result":{"status":"ok","ts":1234567890029,"user":{"k":"kex1p0h0t20x08n28cf5lcncx7llxtrukh6agn4qn09su4pt444ycrxq2x9f8c","n":"name14","sq":1,"sr":"github","u":"https://gist.github.com/name14/1"},"vts":1234567890030}}
+/user/alice@twitter {"kid":"kex1a4yj333g68pvd6hfqvufqkv4vy54jfe6t33ljd3kc9rpfty8xlgs2u3qxr","result":{"status":"ok","ts":1234567890043,"user":{"k":"kex1a4yj333g68pvd6hfqvufqkv4vy54jfe6t33ljd3kc9rpfty8xlgs2u3qxr","n":"alice","sq":1,"sr":"twitter","u":"https://twitter.com/alice/status/1"},"vts":1234567890043}}
+/user/name10@github {"kid":"kex1gwnjuu2yq9mzmantdrpxm77ly6p24myly36wefrp8epy5ra6l57qj97xgz","result":{"status":"ok","ts":1234567890005,"user":{"k":"kex1gwnjuu2yq9mzmantdrpxm77ly6p24myly36wefrp8epy5ra6l57qj97xgz","n":"name10","sq":1,"sr":"github","u":"https://gist.github.com/name10/1"},"vts":1234567890005}}
+/user/name11@github {"kid":"kex1v6l8uvev0fznxv4an5987lds2h6utmc6q6k6vmvckw0mdqgvguaq4850kc","result":{"status":"ok","ts":1234567890010,"user":{"k":"kex1v6l8uvev0fznxv4an5987lds2h6utmc6q6k6vmvckw0mdqgvguaq4850kc","n":"name11","sq":1,"sr":"github","u":"https://gist.github.com/name11/1"},"vts":1234567890010}}
+/user/name12@github {"kid":"kex1pdgn4kd5jfqptjsfqtks0yzy6wk9m0kzxphsd9yvzrdgadhrnuksyklezm","result":{"status":"ok","ts":1234567890015,"user":{"k":"kex1pdgn4kd5jfqptjsfqtks0yzy6wk9m0kzxphsd9yvzrdgadhrnuksyklezm","n":"name12","sq":1,"sr":"github","u":"https://gist.github.com/name12/1"},"vts":1234567890015}}
+/user/name13@github {"kid":"kex1jx3g5zm58q2e8fxeg62hjgyfy6hu3tvzezpekajyxkdeaw56fvaq46atcz","result":{"status":"ok","ts":1234567890020,"user":{"k":"kex1jx3g5zm58q2e8fxeg62hjgyfy6hu3tvzezpekajyxkdeaw56fvaq46atcz","n":"name13","sq":1,"sr":"github","u":"https://gist.github.com/name13/1"},"vts":1234567890020}}
+/user/name14@github {"kid":"kex1p0h0t20x08n28cf5lcncx7llxtrukh6agn4qn09su4pt444ycrxq2x9f8c","result":{"status":"ok","ts":1234567890025,"user":{"k":"kex1p0h0t20x08n28cf5lcncx7llxtrukh6agn4qn09su4pt444ycrxq2x9f8c","n":"name14","sq":1,"sr":"github","u":"https://gist.github.com/name14/1"},"vts":1234567890025}}
