+++ user/store.go
@@ -1,7 +1,6 @@
 package user
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -9,8 +8,6 @@ import (
 
 	"github.com/keys-pub/keys"
 	"github.com/keys-pub/keys/ds"
-	"github.com/keys-pub/keys/encoding"
-	"github.com/keys-pub/keys/link"
 	"github.com/keys-pub/keys/util"
 	"github.com/pkg/errors"
 )
@@ -109,9 +106,9 @@ func (u *Store) Update(ctx context.Context, kid keys.ID) (*Result, error) {
 	return result, nil
 }
 
-// CheckSigchain looks for user in a Sigchain and updates the current result.
+// CheckSigchain looks for user in a Sigchain and updates the current result in the Store.
 func (u *Store) CheckSigchain(ctx context.Context, sc *keys.Sigchain) (*Result, error) {
-	usr, err := FindUserInSigchain(sc)
+	usr, err := FindInSigchain(sc)
 	if err != nil {
 		return nil, err
 	}
@@ -128,117 +125,22 @@ func (u *Store) CheckSigchain(ctx context.Context, sc *keys.Sigchain) (*Result, 
 		}
 	}
 
-	u.updateResult(ctx, usr, result, sc.KID())
+	if usr.KID != sc.KID() {
+		return nil, errors.Errorf("user sigchain kid mismatch %s != %s", usr.KID, sc.KID())
+	}
+
+	updateResult(ctx, u.req, usr, result, u.Now())
 
 	return result, nil
 }
 
-// Check a user. Doesn't index result.
-func (u *Store) Check(ctx context.Context, usr *User, kid keys.ID) (*Result, error) {
-	res := &Result{
-		User: usr,
-	}
-	u.updateResult(ctx, usr, res, kid)
-	return res, nil
+// RequestAndVerify a user. Doesn't index result.
+func (u *Store) RequestAndVerify(ctx context.Context, usr *User) *Result {
+	return RequestAndVerify(ctx, u.req, usr, u.Now())
 }
 
-func userEqual(usr1 *User, usr2 *User) bool {
-	b1, err := json.Marshal(usr1)
-	if err != nil {
-		panic(err)
-	}
-	b2, err := json.Marshal(usr2)
-	if err != nil {
-		panic(err)
-	}
-	return bytes.Equal(b1, b2)
-}
-
-// updateResult updates the specified result.
-func (u *Store) updateResult(ctx context.Context, usr *User, result *Result, kid keys.ID) {
-	if result == nil {
-		panic("no user result specified")
-	}
-	logger.Infof("Update user %s", result.User.String())
-
-	if !userEqual(usr, result.User) {
-		result.Err = "results user is invalid"
-		result.Status = StatusFailure
-		return
-	}
-
-	service, err := link.NewService(result.User.Service)
-	if err != nil {
-		result.Err = err.Error()
-		result.Status = StatusFailure
-		return
-	}
-
-	logger.Debugf("Validate user name: %s, url: %s", result.User.Name, result.User.URL)
-	urs, err := service.ValidateURLString(result.User.Name, result.User.URL)
-	if err != nil {
-		result.Err = err.Error()
-		result.Status = StatusFailure
-		return
-	}
-
-	result.Timestamp = util.TimeToMillis(u.Now())
-
-	logger.Infof("Requesting %s", urs)
-	body, err := u.req.RequestURLString(ctx, urs)
-	if err != nil {
-		logger.Warningf("Request failed: %v", err)
-		if errHTTP, ok := errors.Cause(err).(util.ErrHTTP); ok && errHTTP.StatusCode == 404 {
-			result.Err = err.Error()
-			result.Status = StatusResourceNotFound
-			return
-		}
-		result.Err = err.Error()
-		result.Status = StatusConnFailure
-		return
-	}
-
-	b, err := service.CheckContent(result.User.Name, body)
-	if err != nil {
-		logger.Warningf("Failed to check content: %s", err)
-		result.Err = err.Error()
-		result.Status = StatusContentInvalid
-		return
-	}
-
-	st, err := VerifyContent(b, result, kid)
-	if err != nil {
-		logger.Warningf("Failed to verify content: %s", err)
-		result.Err = err.Error()
-		result.Status = st
-		return
-	}
-
-	logger.Infof("Verified %s", result.User.KID)
-	result.Err = ""
-	result.Status = StatusOK
-	result.VerifiedAt = util.TimeToMillis(u.Now())
-}
-
-// VerifyContent checks content.
-func VerifyContent(b []byte, result *Result, kid keys.ID) (Status, error) {
-	msg, _ := encoding.FindSaltpack(string(b), true)
-	if msg == "" {
-		logger.Warningf("User statement content not found")
-		return StatusContentNotFound, errors.Errorf("user signed message content not found")
-	}
-
-	verifyMsg := fmt.Sprintf("BEGIN MESSAGE.\n%s\nEND MESSAGE.", msg)
-	if _, err := Verify(verifyMsg, kid, result.User); err != nil {
-		logger.Warningf("Failed to verify statement: %s", err)
-		return StatusStatementInvalid, err
-	}
-
-	return StatusOK, nil
-}
-
-// ValidateUserStatement returns error if statement is not a valid user statement.
-func ValidateUserStatement(st *keys.Statement) error {
+// ValidateStatement returns error if statement is not a valid user statement.
+func ValidateStatement(st *keys.Statement) error {
 	if st.Type != "user" {
 		return errors.Errorf("invalid user statement: %s != %s", st.Type, "user")
 	}
@@ -246,7 +148,7 @@ func ValidateUserStatement(st *keys.Statement) error {
 	if err := json.Unmarshal(st.Data, &user); err != nil {
 		return err
 	}
-	if err := ValidateUser(&user); err != nil {
+	if err := Validate(&user); err != nil {
 		return err
 	}
 	return nil
@@ -448,7 +350,7 @@ func (u *Store) Expired(ctx context.Context, dt time.Duration) ([]keys.ID, error
 // CheckForExisting returns key ID of exsiting user in sigchain different from this
 // sigchain key.
 func (u *Store) CheckForExisting(ctx context.Context, sc *keys.Sigchain) (keys.ID, error) {
-	usr, err := FindUserInSigchain(sc)
+	usr, err := FindInSigchain(sc)
 	if err != nil {
 		return "", err
 	}
